add book controller and services, refactor
