+++ src/main/java/dev/almasabdykadyr/library/controller/RentalApi.java
@@ -1,13 +1,9 @@
 package dev.almasabdykadyr.library.controller;
 
 import dev.almasabdykadyr.library.dto.AuthorRequest;
-import dev.almasabdykadyr.library.dto.BookRequest;
 import dev.almasabdykadyr.library.dto.NewRentalRequest;
-import dev.almasabdykadyr.library.dto.UserRequest;
 import dev.almasabdykadyr.library.entity.Author;
-import dev.almasabdykadyr.library.entity.Book;
 import dev.almasabdykadyr.library.entity.Rental;
-import dev.almasabdykadyr.library.entity.User;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.media.ArraySchema;
 import io.swagger.v3.oas.annotations.media.Content;
@@ -23,29 +19,6 @@ import java.util.List;
 @RequestMapping("api/v1")
 @Tag(name = "Rental API", description = "API for Rental Service")
 public interface RentalApi {
-
-    @Operation(
-            summary = "Add a new user",
-            description = "Creates a new user with the provided details.",
-            responses = {
-                    @ApiResponse(responseCode = "201", description = "User successfully created",
-                            content = @Content(mediaType = "application/json", schema = @Schema(implementation = User.class))),
-                    @ApiResponse(responseCode = "400", description = "Invalid input")
-            }
-    )
-    @PostMapping(value = "/users", consumes = "application/json", produces = "application/json")
-    ResponseEntity<User> addUser(@RequestBody @Valid UserRequest request);
-
-    @Operation(
-            summary = "Get all users",
-            description = "Retrieves a list of all users.",
-            responses = {
-                    @ApiResponse(responseCode = "200", description = "Successful operation",
-                            content = @Content(mediaType = "application/json", array = @ArraySchema(schema = @Schema(implementation = User.class))))
-            }
-    )
-    @GetMapping("/users")
-    ResponseEntity<List<User>> getAllUsers();
 
     @Operation(
             summary = "Add a new author",
@@ -70,28 +43,6 @@ public interface RentalApi {
     @GetMapping("/authors")
     ResponseEntity<List<Author>> getAllAuthors();
 
-    @Operation(
-            summary = "Add a new book",
-            description = "Creates a new book with the provided details.",
-            responses = {
-                    @ApiResponse(responseCode = "201", description = "Book successfully created",
-                            content = @Content(mediaType = "application/json", schema = @Schema(implementation = Book.class))),
-                    @ApiResponse(responseCode = "400", description = "Invalid input")
-            }
-    )
-    @PostMapping("/books")
-    ResponseEntity<Book> addBook(@RequestBody @Valid BookRequest request);
-
-    @Operation(
-            summary = "Get all books",
-            description = "Retrieves a list of all books.",
-            responses = {
-                    @ApiResponse(responseCode = "200", description = "Successful operation",
-                            content = @Content(mediaType = "application/json", array = @ArraySchema(schema = @Schema(implementation = Book.class))))
-            }
-    )
-    @GetMapping("/books")
-    ResponseEntity<List<Book>> getAllBooks();
 
     @Operation(
             summary = "Rent a book",
